Support for saving pointers to fields in the state package.
Previously, it was not possible to encode/decode an object graph which
contained a pointer to a field within another type. This was because the
encoder used to track and referenc pointer values by a single start address for
the underyling objects. This cannot disamgibuate between a pointer to the
object and a pointer within the object.

This CL remedies this by constructing an address map tracking the full memory
range object occupy. The encoded Refvalue message has been extended to allow
references to children objects within another object.

See README.md for a description of how the new mechanism works.

This CL also changes the state package to use standard interfaces for
dispatching to Save & Load functions in order to improve performance.

PiperOrigin-RevId: 316163420